Fix tests, implement (ugly) copying back into data directory.
