+++ setup.cfg
@@ -51,8 +51,6 @@ pre_commit =
   pylint==2.6.0
   mypy==0.740
 
-
 [options.entry_points]
 console_scripts =
   aiida-mock-code = aiida_testing.mock_code._cli:run
-
