+++ tests/mock_code/data/mock-diff-broken-16fd960a5a26cfc51d24655b2db7834f/patch.diff
